# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: OpenLoadBalancer Admin API
  description: |
    REST API for managing OpenLoadBalancer instances.

    ## Authentication
    All endpoints require Bearer token authentication via the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting
    API requests are rate-limited to 1000 requests per minute per IP.
  version: 0.1.0
  contact:
    name: OpenLoadBalancer Team
    url: https://openloadbalancer.dev
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8081
    description: Local development
  - url: https://api.openloadbalancer.dev
    description: Production
...
paths:
...
  /api/v1/system/info:
    get:
      tags: [System]
      summary: System information
      description: Returns version, build, uptime, and runtime information
      operationId: getSystemInfo
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: System information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemInfoResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/system/reload:
    post:
      tags: [System]
      summary: Reload configuration
      description: Triggers a hot configuration reload (protected by circuit breaker)
      operationId: reloadConfig
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: Reload succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '429':
          description: Reload already in progress or circuit breaker open
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/routes:
    get:
      tags: [Routes]
      summary: List routes
      description: Returns all configured routing rules
      operationId: listRoutes
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: List of routes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ---------------------------------------------------------------------------
  # Health
  # ---------------------------------------------------------------------------
...
  /api/v1/config:
    get:
      tags: [Config]
      summary: Get configuration
      description: Returns the current live configuration
      operationId: getConfig
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: Current configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [Config]
      summary: Update configuration
      description: Replaces the live configuration (Raft propose in clustered mode, reload in standalone)
      operationId: updateConfig
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Full configuration object matching the YAML config schema
      responses:
        '200':
          description: Configuration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ---------------------------------------------------------------------------
  # Certificates
  # ---------------------------------------------------------------------------
  /api/v1/certificates:
    get:
      tags: [Certificates]
      summary: List certificates
      description: Returns loaded TLS certificates with names, expiry timestamps, and wildcard flags
      operationId: listCertificates
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: Certificate list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificateListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ---------------------------------------------------------------------------
  # WAF
  # ---------------------------------------------------------------------------
  /api/v1/waf/status:
    get:
      tags: [WAF]
      summary: WAF status
      description: Returns Web Application Firewall module status and configuration
      operationId: getWAFStatus
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: WAF status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WAFStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ---------------------------------------------------------------------------
  # Middleware
  # ---------------------------------------------------------------------------
  /api/v1/middleware/status:
    get:
      tags: [Middleware]
      summary: Middleware status
      description: Returns the status of all middleware components
      operationId: getMiddlewareStatus
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: Middleware status list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MiddlewareStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ---------------------------------------------------------------------------
  # Events
  # ---------------------------------------------------------------------------
  /api/v1/events:
    get:
      tags: [Events]
      summary: Activity feed
      description: Returns a recent activity feed of backend health events and system events
      operationId: getEvents
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: Event list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Unauthorized — missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string

    # -------------------------------------------------------------------------
    # System
    # -------------------------------------------------------------------------
    SystemInfo:
      type: object
      properties:
        version:
          type: string
          example: "0.1.0"
        commit:
          type: string
          example: "abc1234"
        build_date:
          type: string
          example: "2024-01-15T10:30:00Z"
        uptime:
          type: string
          example: "2h30m15s"
        state:
          type: string
          enum: [running, stopping, reloading]
        go_version:
          type: string
          example: "go1.26.0"
    SystemInfoResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/SystemInfo'
...
    RouteInfo:
      type: object
      properties:
        name:
          type: string
        host:
          type: string
        path:
          type: string
        methods:
          type: array
          items:
            type: string
        headers:
          type: object
          additionalProperties:
            type: string
        backend_pool:
          type: string
        priority:
          type: integer
    RouteListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/RouteInfo'

    # -------------------------------------------------------------------------
    # Health Checks
    # -------------------------------------------------------------------------
...
    ConfigResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              additionalProperties: true
              description: Full configuration object matching the YAML config schema

    # -------------------------------------------------------------------------
    # Certificates
    # -------------------------------------------------------------------------
    CertificateInfo:
      type: object
      properties:
        names:
          type: array
          items:
            type: string
        expiry:
          type: integer
          format: int64
          description: Expiry as Unix timestamp
        is_wildcard:
          type: boolean
    CertificateListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/CertificateInfo'

    # -------------------------------------------------------------------------
    # WAF
    # -------------------------------------------------------------------------
    WAFStatus:
      type: object
      properties:
        enabled:
          type: boolean
        mode:
          type: string
          enum: [enforce, monitor, disabled]
        ip_acl:
          type: object
          additionalProperties: true
        rate_limit:
          type: object
          additionalProperties: true
        sanitizer:
          type: object
          additionalProperties: true
        detection:
          type: object
          additionalProperties: true
        bot_detection:
          type: object
          additionalProperties: true
        response:
          type: object
          additionalProperties: true
    WAFStatusResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/WAFStatus'

    # -------------------------------------------------------------------------
    # Middleware
    # -------------------------------------------------------------------------
    MiddlewareStatusItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        enabled:
          type: boolean
        category:
          type: string
    MiddlewareStatusResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/MiddlewareStatusItem'

    # -------------------------------------------------------------------------
    # Events
    # -------------------------------------------------------------------------
    EventItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [success, info, warning, error]
        message:
          type: string
        timestamp:
          type: string
    EventListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/EventItem'

    # -------------------------------------------------------------------------
    # Cluster
    # -------------------------------------------------------------------------
...
security:
  - bearerAuth: []